backend/requirements.txt
python-docx>=0.8.11
python-dateutil>=2.8.2
openpyxl>=3.1.5
